refactor: move FormalForkTester
